requirements/vision.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cu128 requirements/vision.in
#
--extra-index-url https://download.pytorch.org/whl/cu128

certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
filelock==3.20.1
    # via torch
fsspec==2025.12.0
    # via torch
idna==3.11
    # via requests
jinja2==3.1.6
    # via torch
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.2.6
    # via
    #   opencv-contrib-python
    #   torchvision
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
opencv-contrib-python==4.12.0.88
    # via -r requirements/vision.in
packaging==25.0
    # via pytesseract
pillow==12.0.0
    # via
    #   pytesseract
    #   torchvision
pytesseract==0.3.13
    # via -r requirements/vision.in
requests==2.32.5
    # via -r requirements/base.in
sympy==1.14.0
    # via torch
torch==2.9.1+cu128
    # via
    #   -r requirements/vision.in
    #   torchvision
torchvision==0.24.1+cu128
    # via -r requirements/vision.in
triton==3.5.1
    # via torch
typing-extensions==4.15.0
    # via torch
urllib3==2.6.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
